cmd/govim: highlight diagnostics
Currently diagnostics as errors and warnings are only placed
in the quickfix buffer, with a sign in the gutter on that line.
This change also adds text property highlighting directly to
the range of the code affected by the diagnostic.

Note that this is a breaking change if you have overridden
any of the default govim highlight rules to, for example,
change the color of error signs. Update your overrides according
to the new highlight names.

Override the following highlights in your .vimrc to alter the
appearance of text properties and signs.
- govimSignErr, govimSignWarn, govimSignInfo & govimSignHint
- govimErr, govimWarn, govimInfo & govimHint